first_ai_chatbot/requirements.txt
streamlit>=1.28.0
google-genai>=1.24.0